+++ ostap/parallel/parallel_tmva.py
@@ -105,7 +105,18 @@ def addTMVAResponse ( chain                  ,   ## input chain
     >>> dataset.addTMVAResponse (  inputs , tar_file , prefix = 'tmva_' )
     """
     from ostap.tools.tmva import addTMVAResponse as _add_response_
+
+    assert prefix or suffix , 'addTMVAResponse: invalid prefix/suffix %s/%s' % ( prefix , suffix ) 
     
+    if isinstance ( chain , ROOT.TTree ) :
+        import ostap.trees.trees        
+        vars    = set ( chain.branches() ) | set ( chain.leaves () ) 
+        matched = sorted ( v for v in vars if v.startswith ( prefix ) and v.endswith (  suffix ) ) 
+        if matched :
+            matched = ','.join ( matched ) 
+            logger.warning ( "addTMVAResponse:: Variables '%s' already in TTree, skip" % matched ) 
+            return chain
+        
     if isinstance ( chain , ROOT.TChain ) and 1 < len ( chain.files () ) : pass
     else : return _add_response_ ( dataset       = chain         ,
                                    inputs        = inputs        ,
@@ -117,7 +128,7 @@ def addTMVAResponse ( chain                  ,   ## input chain
 
     from ostap.trees.trees import Chain
     ch       = Chain ( chain ) 
-    branches = set   ( chain.branches() )
+    branches = set   ( chain.branches() ) | set ( chain.leaves() ) 
     
     ## create the task 
     task  = AddTMVA     ( inputs        = inputs        ,
@@ -135,7 +146,7 @@ def addTMVAResponse ( chain                  ,   ## input chain
     nc = ROOT.TChain ( chain.name )
     for f in ch.files :  nc.Add ( f )
 
-    nb = list ( set ( nc.branches () ) - branches )
+    nb = list ( ( set ( nc.branches () ) | set ( nc.leaves() ) ) - branches )
     if nb : logger.info ( 'Added branches:\n%s' % nc.table ( variables = nb , prefix = '# ' ) ) 
     
     return nc